CU-2hqn0ec - Reverted back to using headless UI tabs as intended rather than links to different routes
