+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+  "workbench.tree.indent": 20
+}
